requirements.txt
redis
google-api-python-client
psycopg2-binary
streamlit
pandas
